+++ src/Components/Frontpage/SuccessStories/SuccessStories.module.scss
@@ -1,0 +1,37 @@
+.SuccessStories {
+  background: #e3edf3;
+  text-align: center;
+  padding: 50px 0 50px;
+  h2 {
+    color: #425762;
+    margin: 0 0 12px;
+    font-size: 36px;
+    line-height: 40px;
+  }
+  .item {
+    margin: 2% 2% 0 0;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.25);
+    border: 1px solid #b5cedd;
+    background: #fff;
+    .title {
+      font-size: 24px;
+      line-height: 26px;
+      display: block;
+      margin-bottom: 12px;
+      font-weight: 500;
+    }
+    .img-block {
+      width: 100%;
+      background-size: cover;
+      height: 320px;
+      margin: 0 0 0 4px;
+    }
+    .content-area {
+      padding: 14px 20px 10px;
+      overflow: hidden;
+      p {
+        margin: 0 0 40px;
+      }
+    }
+  }
+}
