+++ src/Components/Frontpage/SuccessStories/SuccessStoriesState.tsx
@@ -1,0 +1,5 @@
+import { Suggestion } from "../../Common/Suggestion";
+export interface SuccessStoriesState {
+	suggestions?: Array<Suggestion>;
+	CurrentImage?: number;
+}
